chore(data-workflows): Rename RunsItem => ItemComponent
